WIP: feature put_space DONE
